requirements.txt
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
docutils==0.21.2
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
packaging==24.2
pkginfo==1.10.0
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dotenv==1.0.1
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
setuptools==75.6.0
tenacity==9.0.0
twine==5.1.1
urllib3==2.2.3
zipp==3.21.0
